enum-support: in Transfer, use base ValueEnumBase when reading
